+++ lib/src/ui/page/photo_main_page.dart
@@ -65,6 +65,8 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   String get currentGalleryName {
     if (currentPath?.isAll == true) {
       return i18nProvider.getAllGalleryText(options);
+    } else if (currentPath == null) {
+      return i18nProvider.getNoSelectedText(options);
     }
     return currentPath?.name ?? "Select Folder";
   }
@@ -81,12 +83,21 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   @override
   void initState() {
     super.initState();
-    _refreshList();
     scaffoldKey = GlobalKey();
     scrollController = ScrollController();
     _changeThrottle = Throttle(onCall: _onAssetChange);
     PhotoManager.addChangeCallback(_changeThrottle.call);
     PhotoManager.startChangeNotify();
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (!_isInit) {
+      final pickedList = PhotoPickerProvider.of(context).pickedAssetList ?? [];
+      addPickedAsset(pickedList.toList());
+      _refreshList();
+    }
   }
 
   @override
@@ -189,7 +200,8 @@ class _PhotoMainPageState extends State<PhotoMainPage>
     );
   }
 
-  void _refreshList() {
+  void _refreshList() async {
+    await Future.delayed(Duration.zero);
     if (!useAlbum) {
       _refreshListFromWidget();
       return;
@@ -199,30 +211,27 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   }
 
   Future<void> _refreshListFromWidget() async {
-    galleryPathList.clear();
-    galleryPathList.addAll(widget.photoList);
-    this.list.clear();
-    var assetList = await galleryPathList[0].assetList;
-    _sortAssetList(assetList);
-    this.list.addAll(assetList);
-    setState(() {
-      _isInit = true;
-    });
+    _onRefreshAssetPathList(widget.photoList);
   }
 
   Future<void> _refreshListFromGallery() async {
     List<AssetPathEntity> pathList;
     switch (options.pickType) {
       case PickType.onlyImage:
-        pathList = await PhotoManager.getImageAsset();
+        pathList = await PhotoManager.getAssetPathList(type: RequestType.image);
         break;
       case PickType.onlyVideo:
-        pathList = await PhotoManager.getVideoAsset();
+        pathList = await PhotoManager.getAssetPathList(type: RequestType.video);
         break;
       default:
-        pathList = await PhotoManager.getAssetPathList();
+        pathList = await PhotoManager.getAssetPathList(
+            type: RequestType.image | RequestType.video);
     }
 
+    _onRefreshAssetPathList(pathList);
+  }
+
+  Future<void> _onRefreshAssetPathList(List<AssetPathEntity> pathList) async {
     if (pathList == null) {
       return;
     }
@@ -246,10 +255,6 @@ class _PhotoMainPageState extends State<PhotoMainPage>
     setState(() {
       _isInit = true;
     });
-  }
-
-  void _sortAssetList(List<AssetEntity> assetList) {
-    options?.sortDelegate?.assetDelegate?.sort(assetList);
   }
 
   Widget _buildBody() {
@@ -396,7 +401,7 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   }
 
   void _onItemClick(AssetEntity data, int index) {
-    var result = new PhotoPreviewResult();
+    var result = PhotoPreviewResult();
     isPushed = true;
     Navigator.of(context).push(
       MaterialPageRoute(
@@ -427,7 +432,7 @@ class _PhotoMainPageState extends State<PhotoMainPage>
   }
 
   void _onTapPreview() async {
-    var result = new PhotoPreviewResult();
+    var result = PhotoPreviewResult();
     isPushed = true;
     var v = await Navigator.of(context).push(
       MaterialPageRoute(
@@ -502,10 +507,10 @@ class _PhotoMainPageState extends State<PhotoMainPage>
     List<AssetPathEntity> pathList;
     switch (options.pickType) {
       case PickType.onlyImage:
-        pathList = await PhotoManager.getImageAsset();
+        pathList = await PhotoManager.getAssetPathList(type: RequestType.image);
         break;
       case PickType.onlyVideo:
-        pathList = await PhotoManager.getVideoAsset();
+        pathList = await PhotoManager.getAssetPathList(type: RequestType.image);
         break;
       default:
         pathList = await PhotoManager.getAssetPathList();